+++ gdsfactory/component_reference.py
@@ -1,12 +1,11 @@
 import typing
-from collections import defaultdict
 from typing import Any, Dict, List, Optional, Tuple, Union, cast
 
 import gdstk
 import numpy as np
 from numpy import cos, float64, int64, mod, ndarray, pi, sin
 
-from gdsfactory.component_layout import _GeometryHelper
+from gdsfactory.component_layout import Polygon, _GeometryHelper, get_polygons
 from gdsfactory.port import (
     Port,
     map_ports_layer_to_orientation,
@@ -221,11 +220,11 @@ class ComponentReference(_GeometryHelper):
 
     def get_polygons(
         self,
-        by_spec=False,
-        depth=None,
+        by_spec: bool = False,
+        depth: Optional[int] = None,
         include_paths: bool = True,
         as_array: bool = True,
-    ):
+    ) -> Union[List[Polygon], Dict[Tuple[int, int], List[Polygon]]]:
         """Return the list of polygons created by this reference.
 
         Args:
@@ -252,39 +251,13 @@ class ComponentReference(_GeometryHelper):
             Instances of `FlexPath` and `RobustPath` are also included in
             the result by computing their polygonal boundary.
         """
-        if not by_spec:
-            polygons = self._reference.get_polygons(
-                depth=depth, include_paths=include_paths
-            )
-        elif by_spec is True:
-            layers = self.parent.get_layers()
-            polygons = {
-                layer: self._reference.get_polygons(
-                    depth=depth,
-                    layer=layer[0],
-                    datatype=layer[1],
-                    include_paths=include_paths,
-                )
-                for layer in layers
-            }
-
-        else:
-            polygons = self._reference.get_polygons(
-                depth=depth,
-                layer=by_spec[0],
-                datatype=by_spec[1],
-                include_paths=include_paths,
-            )
-
-        if not as_array:
-            return polygons
-        if not by_spec:
-            return [polygon.points for polygon in polygons]
-        layer_to_polygons = defaultdict(list)
-        for layer, polygons_list in polygons.items():
-            for polygon in polygons_list:
-                layer_to_polygons[layer].append(polygon.points)
-        return layer_to_polygons
+        return get_polygons(
+            instance=self,
+            by_spec=by_spec,
+            depth=depth,
+            include_paths=include_paths,
+            as_array=as_array,
+        )
 
     def get_labels(self, depth=None, set_transform=True):
         """Return the list of labels created by this reference.
@@ -821,8 +794,47 @@ def test_move():
     bend.move("o1", mzi.ports["o2"])
 
 
-if __name__ == "__main__":
+def test_get_polygons():
     import gdsfactory as gf
+
+    ref = gf.components.straight()
+    p0 = ref.get_polygons(by_spec="WG", as_array=False)
+    p1 = ref.get_polygons(by_spec=(1, 0), as_array=True)
+    p2 = ref.get_polygons(by_spec=(1, 0), as_array=False)
+
+    p3 = ref.get_polygons(by_spec=True, as_array=True)[(1, 0)]
+    p4 = ref.get_polygons(by_spec=True, as_array=False)[(1, 0)]
+
+    assert len(p1) == len(p2) == len(p3) == len(p4) == 1 == len(p0)
+    assert p1[0].dtype == p3[0].dtype == float
+    assert isinstance(p2[0], Polygon)
+    assert isinstance(p4[0], Polygon)
+
+
+def test_get_polygons_ref():
+    import gdsfactory as gf
+
+    ref = gf.components.straight().ref()
+    p0 = ref.get_polygons(by_spec="WG", as_array=False)
+    p1 = ref.get_polygons(by_spec=(1, 0), as_array=True)
+    p2 = ref.get_polygons(by_spec=(1, 0), as_array=False)
+
+    p3 = ref.get_polygons(by_spec=True, as_array=True)[(1, 0)]
+    p4 = ref.get_polygons(by_spec=True, as_array=False)[(1, 0)]
+
+    assert len(p1) == len(p2) == len(p3) == len(p4) == 1 == len(p0)
+    assert p1[0].dtype == p3[0].dtype == float
+    assert isinstance(p2[0], Polygon)
+    assert isinstance(p4[0], Polygon)
+
+
+if __name__ == "__main__":
+    test_get_polygons_ref()
+    test_get_polygons()
+    import gdsfactory as gf
+
+    ref = gf.components.straight().ref()
+    p = ref.get_polygons(by_spec=(1, 0), as_array=False)
 
     # c = gf.Component("parent")
     # c2 = gf.Component("child")
@@ -832,8 +844,8 @@ if __name__ == "__main__":
     # c2.add_polygon([(0, 0), (length, 0), (length, width), (0, width)], layer=layer)
     # c << c2
 
-    c = gf.c.dbr()
-    c.show()
+    # c = gf.c.dbr()
+    # c.show()
 
     # import gdsfactory as gf