+++ gdsfactory/font.py
@@ -203,3 +203,10 @@ def _get_glyph(font, letter):  # noqa: C901
     # Cache the return value and return it
     font.gds_glyphs[letter] = (component, glyph.advance.x / font.size.ascender)
     return font.gds_glyphs[letter]
+
+
+if __name__ == "__main__":
+    from gdsfactory.components.text_freetype import text_freetype
+
+    c = text_freetype("hello", font="Times New Roman")
+    c.show(show_ports=True)